+++ framework/domain/src/main/java/is/codion/framework/domain/entity/condition/DefaultConditionCombination.java
@@ -61,17 +61,17 @@ final class DefaultConditionCombination extends AbstractCondition implements Com
 	}
 
 	@Override
-	public String toString(EntityDefinition definition) {
+	public String string(EntityDefinition definition) {
 		requireNonNull(definition);
 		if (conditions.isEmpty()) {
 			return "";
 		}
 		if (conditions.size() == 1) {
-			return conditions.get(0).toString(definition);
+			return conditions.get(0).string(definition);
 		}
 
 		return conditions.stream()
-						.map(condition -> condition.toString(definition))
+						.map(condition -> condition.string(definition))
 						.filter(string -> !string.isEmpty())
 						.collect(joining(toString(conjunction), "(", ")"));
 	}