+++ changelog.md
@@ -6,6 +6,7 @@ Codion Change Log
 - FilterTable.Builder.hiddenColumns() added.
 ### is.codion.framework.domain
 - DerivedAttributeDefinition.Builder.SourceAttributesStep added.
+- Condition.toString(entityDefinition) renamed string() in order to not be confused with toString().
 ### is.codion.swing.framework.ui
 - EntitySearchFieldPanel.Builder.preferredSearchFieldWidth() added.
 - EntityPanel.createMainComponent() bug fixed, mainPanel() got called twice if a detail panel was present, messing up the controls panel, now supplier based.